requirements.txt
pillow
psycopg2
